+++ src/ascnhalf-world/Container.h
@@ -33,9 +33,9 @@ public:
 	void Create( uint32 itemid, Player* owner );
 	void LoadFromDB( Field*fields);
 
-	bool AddItem(int8 slot, Item* item);
+	bool AddItem(int16 slot, Item* item);
 	bool AddItemToFreeSlot(Item* pItem, uint32 * r_slot);
-	Item* GetItem(int8 slot)
+	Item* GetItem(int16 slot)
 	{
 		if(slot >= 0 && (uint8)slot < GetProto()->ContainerSlots)
 			return m_Slot[slot];
@@ -46,11 +46,11 @@ public:
 	int8 FindFreeSlot();
 	bool HasItems();
 	
-	void SwapItems(int8 SrcSlot,int8 DstSlot);
-	Item* SafeRemoveAndRetreiveItemFromSlot(int8 slot, bool destroy); //doesnt destroy item from memory
-	bool SafeFullRemoveItemFromSlot(int8 slot); //destroys item fully
+	void SwapItems(int16 SrcSlot, int16 DstSlot);
+	Item* SafeRemoveAndRetreiveItemFromSlot(int16 slot, bool destroy); //doesnt destroy item from memory
+	bool SafeFullRemoveItemFromSlot(int16 slot); //destroys item fully
    
-	void SaveBagToDB(int8 slot, bool first, QueryBuffer * buf);
+	void SaveBagToDB(int16 slot, bool first, QueryBuffer * buf);
 
 protected:
 	Item* m_Slot[72];
